doc(test): suggest --no-build in conjunction with --ab
When running A/B tests, this part of the documentation explicitly builds
firecracker from two revisions. However, after that it will _still_
build Firecracker from HEAD _again_, even though this third compilation
of firecracker is not needed: The A/B test will only ever use the
precompiled binaries.

Fix this by passing --no-build in the example

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>