fix(test): eliminate false positives from is-microvm-dead check
Ensure that `ps(1)` does not truncate the command, which might result in
the grep failing (if the jailer_id gets truncated), using the -ww
option. While we're at it, also use -o cmd so that ps only prints the
command names and nothing else (as we're not using anything else from
this output).

This causes false-positives instead of false-negatives funnily enough,
because we're using check_output, meaning if the grep doesnt find
anything we fail the command (in the "everything works" scenario,
firecracker is dead but grep still matches the "ps | grep" process
itself).

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>