test(perf): disable memory monitor for memory overhead perf test
This test just boots a VM, which a ton of other tests also do, so if
memory overhead really does change, we'll catch it in other tests. On
the other hand, having this test just crash if memory overhead goes
above 5MB is not very useful, because it prevent this test from being
run as a A/B-test in scenarios where memory overhead is indeed
increasing.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>