test(ab): allow ignoring specific metric with ignore list
Currently, if something matches the A/B-testing ignore list, then all
metrics emitted from a test with a dimension set that is a super set of
an ignored one is ignored. Refine this to allow only ignoring specific
metrics.

Realize this by synthesizing a fake dimension called 'metric' that
stores the metric.

This will later be used when we introduce block latency tests, as we
will want to A/B-test throughput but ignore latency in scenarios where
fio's async workload generator is used.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>