allow specifying custom cpu template inline in config json
When starting firecracker without the API server and configuring it
through a config files, specifying a custom cpu template requires
writing this cpu template into a separate json file, and then passing
the path to this separate json file in the config json file. This is a
bit silly, since everything is just json. So additionally allow just
directly specifying the custom cpu template inside the config file.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>