test(jailer): use tmp dir for mknod test
We seldom have failures in the CI where the test_mknod_and_own_dev fails
because a file already exists.
The test is using the actual /dev to create tmp devices. As there's no
reason to use the actual /dev, move it to use a random folder and clean
it up after the test.

Signed-off-by: Riccardo Mancini <mancio@amazon.com>