test(balloon): ensure we gave stats enough time to update before reading
test_balloon_snapshot started being flaky after we changed the logic on
how we wait for the RSS to become stable.
One theory is that we are not waiting enough time for the stats to
refresh, so this change adds a sleep to ensure we have waited enough for
the stats to be "fresh".

Failure:
```
assert 189022208 > 189022208
```

Signed-off-by: Riccardo Mancini <mancio@amazon.com>