fix(test/pvtime): bump tolerance to 10s
We found a single failure for which the steal time between snapshot
and restore went slightly above 2s on an AMD instance.
As the purpose of this check is to ensure the value is "sane" (iow not
a completely random number), not that it's really accurate (that's a
kernel problem), I'm bumping it to 10s.

Signed-off-by: Riccardo Mancini <mancio@amazon.com>