test(signal): remove flaky unit test
The signal handler unit test registers the signal handlers, spins up a
new thread, and sends signals to itself (using kill) to check that the
metrics get updated. This is a very complicated test for a unit test,
and it's already covered in the integration test test_signals.py.
As this test is seldomly failing in our CI, it's best to get rid of its
complexity and rely on the integration tests, which are better suited
for this kind of test.

Signed-off-by: Riccardo Mancini <mancio@amazon.com>