test: add network interface to test_serial_dos test
test_serial_dos test checks that when we send a lot of bytes in the
serial device the emulation logic does not increase indefinitely the
underlying buffer that we use for when the device is set in loopback
mode.

However, the test does not wait for the microVM to start and sometimes
the virtual memory allocation may increase between readings. Add a
network device to the microVM so that we implicitly wait until it has
booted before taking the first measurement.

Signed-off-by: Babis Chalios <bchalios@amazon.es>