fix(test): split post-snapshot checks for steal time test
test_pvtime_snapshot checks that steal time reporting mechanism works
on snapshotted microVMs. This includes two checks; first that the old
steal time values are saved and restored correctly through snapshots and
second that steal time keeps increasing in microVMs resumed from a
snapshot.

The test fails intermittently, but we are not sure which of the two
invariants causes it to fail because we are checking both in the same
assertion. Split this assertion in two so that we understand better the
culprit.

Signed-off-by: Babis Chalios <bchalios@amazon.es>