refactor(test): spawn uffd handler inside restore_from_snapshot
Drop the requirement of having to call spawn_pf_handler() before the
call to microvm.restore_from_snapshot() in integration tests, and
instead just have restore_from_snapshot() take the name of the uffd
handler as an optional keyword argument, and do the spawning. This
removes the weird calling convention of calling restore_from_snapshot()
without a snapshot parameter (because it was inferred from uffd).

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>