riscv64: add interrupt support for block devices
This commit adds interrupt support for VirtIO block devices on riscv64,
providing the necessary functionality to support interrupts for VirtIO
devices in general. Interrupt delivery is implemented using KVM's
in-kernel irqchip.

Unlike x86_64 and aarch64, we couldn’t get the existing IRQFD-based
approach to work reliably on riscv64. As a result, this implementation
uses the `KVM_IRQ_LINE` ioctl to signal interrupts.

To support this, it extends the `IrqTrigger` structure to hold both the
raw VM file descriptor and the GSI number. Storing the raw file
descriptor was necessary because `IrqTrigger` cannot hold a `VmFd`
directly, as this type doesn't implement `Clone`. Having the raw VM file
descriptor and the corresponding GSI allows `trigger_irq()` to perform
the necessary ioctls to assert and de-assert interrupt lines when
needed.

Signed-off-by: Dimitris Charisis <dchar@cslab.ece.ntua.gr>