riscv64: add interrupt support for serial device
Implement interrupt support for the legacy serial console device on
riscv64, following the same logic as the VirtIO device implementation.
Replace the `EventFdTrigger` structure, which triggers IRQFD-related
interrupts on x86_64 and aarch64, with a new `IrqLineTrigger` type. This
type encapsulates the two required pieces of information to fire a
`KVM_IRQ_LINE` interrupt: the raw file descriptor behind the `VmFd`
type and the corresponding GSI number.

This change is necessary because the `VmFd` structure does not implement
`Clone`, preventing it from being carried into the `trigger()` function
when injecting an interrupt from the serial device.

Additionally, make the `allocate_mmio_resources()` method of
`MMIODeviceManager` public. This is required to allocate an IRQ number
*before* setting up the serial device, since `setup_serial_device()`
builds the `IrqLineTrigger` and needs to know the IRQ number associated
with the serial device at that point.

Signed-off-by: Dimitris Charisis <dchar@cslab.ece.ntua.gr>