mmds: Respond JSON object for credential paths
Some clients send requests with `Accept: application/json` and expects
a JSON object to be returned for the AWS credential paths. On the other
hand, MMDS responds a string of JSON if `Accept: application/json`. This
behavior is expected and the response body is valid as JSON. What such
clients should do is to request with `Accept: text/plain` or without
`Accept` header to get an IMDS-compatible result as documented in our
doc, but we apply a short-term workaround specially for such clients.

Signed-off-by: Takahiro Itazuri <itazur@amazon.com>