refactor(test): add more debugging for "jailer id still used"
Clean up the error messages we print when running into the "firecracker
pid was killed, but process with its jailer id still exists" case: Also
print the PID of the offending process, as well as filter out
non-firecracker processes from the error message (otherwise the
error message always contained the grep process itself, which was
confusing).

Note that `cmd` must be the last column, as otherwise ps will truncate
it.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>