fix(test): do no cause secondary errors in _dump_debug_info
If firecracker already died, then the pid file might not exist anymore,
and we get a NoSuchProcess exception from thread_backtraces(). This can
be confusing when debugging failures, because its a secondary error that
only happens during error handling of the actual reason we're dumping
debug information in the first place.

Fix this by simply returning an empty list of threads in this case.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>