test: cleanup and improve "wait for readiness" checks
In Microvm.spawn(), we try to wait for the firecracker process to
initialize itself and become ready to server requests. We have multiple
checks of varying fidelity in there, and they can be strictly ordered by
this based on what they wait for: e.g. if we wait for SSH availability
(guest userspace is ready), there is no point to _also_ wait for
firecracker's startup message in the logs, as that is always printed
before SSH becomes available. Thus clean this logic up to only do the
one check that has the highest fidelity in any given setup.

While we're at it, update/move some comments.

Then, improve the check for API server readiness to wait for the log
message that signals completion of API server initialization, instead of
just waiting for socket file creation (which happens before we are
actually ready to accept connections on it).

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>