test: avoid filing negative tests due to api server not starting
Some negative tests check for a failure mode that happens before we can
print the "API server started" log message, which then causes issues in
Microvm.spawn(). Work around this by adjusting some except clauses (it
seems that even in the past we ran into this issue due to the old "does
api socket file exist" check), or setting the log level to WARN, which
also disables the "api server started" check. A special case if the
describe snapshot version check, which runs into issues because the "api
server started" log message is fairly new (1.12), and some older
firecrackers do not print it.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>