pci: add virtio-pci transport implementation
Add a VirtIO PCI transport implementation. Nothing uses it at the
moment. This requires a few changes in our vended pci and vm-device
crates.

Add a couple of tests that ensure that PCI configuration space is what
expected. We read common fields and make sure the BAR we allocate for
the VirtIO device is what expected.

Signed-off-by: Babis Chalios <bchalios@amazon.es>