fix: boottimer device MMIO address
When we re-arranged the MMIO address space in commit 9a165d17f1ba (arch:
define 64-bit capable MMIO memory regions) we moved the MMIO region of
the boot timer device for x86 systems, but we didn't update the init
scripts that hardcode it and use it to report boot time timestamp back
to Firecracker.

Update the init.c and initramfs values for the region. Also, add a
functional test that runs during CI PR tests and makes sure the boot
timer works.

Signed-off-by: Babis Chalios <bchalios@amazon.es>