refactor(vm): move ResourceAllocator inside Vm
ResourceAllocator object was part of DeviceManager since it is (mainly)
devices that use it. ResourceAllocator is as well the object that
implements (in a dummy way, for the moment) the DeviceRelocation trait
which PciDevices use to move the address space of a PciDevice when
triggered from the guest.

Problem with DeviceRelocation is that it also needs the Vm file
descriptor to perform the relocation, because we need to move register
the new IO event fd for VirtIO devices.

To make things simpler, move ResourceAllocator inside the Vm object. In
subsequent commit we will remove the DeviceRelocation from
ResourceAllocator and move it to Vm instead.

This has the nice secondary effect that we were able to simplify the
signature of many device-related methods that received Vm and
ResourceAllocator arguments.

Signed-off-by: Babis Chalios <bchalios@amazon.es>