refactor(vm): move `Bus` objects to Vm
We had previously added MMIO and Port IO buses inside ResourceAllocator
so that we could implement DeviceRelocation for the type. Now, we will
delegate device relocation responsibilities to ArchVm instead. That is
because device relocation requires access to the Vm file descriptor as
well.

As a result, we can move buses to the Vm object itself. Add MMIO bus to
VmCommon as both architectures use it. Add PortIO bus for x86
architecture only.

Not that we don't still support DeviceRelocation. VirtIO devices should
not request us to relocate them. Also, for adding such support we would
need to also support VirtIO reset. We will look into adding this
functionaliyt later on.

Signed-off-by: Babis Chalios <bchalios@amazon.es>