test: add a test that checks for starved VirtIO devices
Firecracker IO is software emulated by the VMM thread of the Firecracker
process. This is a single thread so it could be the case that a single
device can starve other devices. This could happen if a guest keeps
adding descriptors in a queue while we are processing it. Our emulation
logic should include railguards against such behaviour. Add a test that
verifies that this is the case.

Signed-off-by: Babis Chalios <bchalios@amazon.es>