test: drop _ab suffix from performance tests
There is nothing that makes these tests specifically "ab" - any test can
be run as an A/B-test as long as it produces more than one data point
for a metric. So drop this suffix. It was only added originally when we
transitioned to A/B-testing and the A/B-capable versions of the tests
had to co-exist with the old baseline version of the tests in separate
files.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>