query correct metric for diff snapshots in snap creation perf test
We have different metrics for snapshot creation latency for full and
diff snapshots (why? Good question), so if we're measuring diff snapshot
creation latency we need to query diff_create_snapshot instead of
full_create_snapshot.

Just inline this logic into the test instead of adding a property to
SnapshotType because there is literally no other use of this pattern in
our tests (if there ever will be, which I doubt, we can still abstract
it away).

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>