feat(virtio-mem): Implement device actions
This patch implements all device actions using a simple bitmap to track
the state of the plugged memory.
No allocation is done on demand, the memory is still all allocated ahead
of time from builder.rs.
There is also no API to trigger the hot(un)plug, so the requested_size
is hardcoded to the total size.

Signed-off-by: Riccardo Mancini <mancio@amazon.com>