fix(mmds): Validate X-metadata-token-ttl-seconds header only if needed
We checked that "X-metadata-token-ttl-seconds" has a valid positive
number regardless of the request type. On the oher hand, EC2 IMDS only
validates it only if it is a PUT request to "/latest/api/token". Such
a behavioral difference from EC2 IMDS could result in unexpected issues.

To line up with EC2 IMDS, breaks `TokenHeaders` down to dedicated
structs (`XMetadataToken` and `XMetadataTokenTtlSeconds`) and parses
"X-metadata-token" only for GET requests and
"X-metadata-token-ttl-seconds" only for PUT requests to
"latest/api/token".

Signed-off-by: Takahiro Itazuri <itazur@amazon.com>