test: extend test_max_devices to support aarch64
The test_max_devices test verifies the maximum number of devices that
can be attached, as permitted by the platform. It checks both that the
expected number of devices can be successfully attached and that
attaching more than the limit results in a failure.

Previously, this test was only executed on x86_64, where 18 devices are
expected. This update extends the test to run on aarch64 as well, which
supports up to 94 devices.

Signed-off-by: Sheng-Wei (Way) Chen <waychensw@gmail.com>