feat(mmds): Support token generation via PUT in MMDS v1
To enable smoother migration from v1 to v2, make MMDS v1 support PUT
request to /latest/api/token for token generation.

We will not make MMDS v1 deny GET requests with invalid token in order
not to break existing workloads. It does not validate a given toekn at
this point, but it will validate to increment a metric that will be
added to count such invalid tokens.

Signed-off-by: Takahiro Itazuri <itazur@amazon.com>