chore: update patch series
Update the patch series from which the CI builds the secret hiding test
kernel. Mainly, replace the v4 [1] of the mmap-support series with the
v12 [2]. Then rebase it on top of 6.16-rc3 to pick per-module exports
that were merged for 6.16.

Update the direct map removal patches to what I locally have written for
the upcoming v5, folding in the hotfix. Add a new patch that makes
kvm-clock work with direct-map removed guest_memfd.

UFFD patches are just rebases, particularly KVM Userfault is still the
v2.

[1]: https://lore.kernel.org/kvm/20250218172500.807733-1-tabba@google.com/
[2]: https://lore.kernel.org/kvm/20250611133330.1514028-1-tabba@google.com/

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>