refactor(metrics): avoid temporary buffer when serializing metrics
Metrics are currently written in a temporary buffer (String) before
being written to the file.
There is no need for this temporary buffer, as the metrics can be
serialized directly into the writer.

I found this as I saw Firecracker was allocating and freeing memory
every minute, causing pages to be mmap/munmap-ed periodically (and
unnecessarily).

Signed-off-by: Riccardo Mancini <mancio@amazon.com>