fix: update base commit to -rc6
Linus fixed an issue introduced between -rc4 and -rc5 that shows up as
random userspace processes hanging without and kernel logs to show what
might be going from [1]. This pretty much exactly matches what we are
seeing with docekr just random hanging sometimes. Let's try updating to
-rc6 to see if that makes the issue go away.

On aarch64, this kernel needs some additional dependencies to build
(hexdump), which we do not have in the docker container, so add a
"instlal build dependencies" step to the kernel building script.

[1]: https://lore.kernel.org/all/CAHk-=wiMJWwgJ4HYsLzJ4_OkhzJ75ah0HrfBBk+W-RGjk4-h2g@mail.gmail.com/

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>