example(uffd): dont panic if read(2) from uffd returns -EAGAIN
Started seeing the below failure in test_population_latency:

thread 'main' panicked at .../uffd/fault_all_handler.rs:41:18:
uffd_msg not ready
note: run with `RUST_BACKTRACE=1` environment variable to display a
backtrace

I am not entierly sure how this can happen, because the read from the
uffd is supposed to be blocking, but maybe it's a weird interaction
with the fault-all behavior (e.g. there was a uffd event queues, but
because we faulted everything it got cancelled again?), so let's just
try going back to read(2) if we dont read anything.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>