refactor(vmm): Remove code for restoring old major version snapshots
The snapshot versioning follows the semantic versioning (i.e. snapshot
restore is allowed only if the restore major version is same as the
snapshot major version and the restore minor version is greater than or
equal to the snapshot minor version). We already bumped the major
version up multiple times since we added `mmds_version` state. That
means snapshots taken when `mmds_version` didn't exist are not able to
be restored in the current version. The code to restore such snapshots
is no longer needed.

Signed-off-by: Takahiro Itazuri <itazur@amazon.com>